Code cleaning

+++ src/main/java/me/codeflusher/gcalc/util/Utils.java
@@ -22,22 +22,23 @@ public class Utils {
         return buffer;
     }
 
-    public static int parseInt(String s){
+    public static int parseInt(String s) {
         StringBuilder builder = new StringBuilder();
         for (char c : s.toCharArray()) {
-            if (Constants.NUM_MAP.getOrDefault(c, -1) != -1){
+            if (Constants.NUM_MAP.getOrDefault(c, -1) != -1) {
                 builder.append(c);
             }
         }
         return Integer.parseInt(builder.toString());
     }
 
-    public static float clampFloat(float value, float minimum, float maximum){
+    public static float clampFloat(float value, float minimum, float maximum) {
         if (value < minimum) return minimum;
         return Math.min(value, maximum);
     }
-    public static int limitSizeOfModel(int inputSize){
-        if (inputSize < 16){
+
+    public static int limitSizeOfModel(int inputSize) {
+        if (inputSize < 16) {
             return 16;
         }
         return Math.min(inputSize, 1536);
@@ -58,7 +59,7 @@ public class Utils {
         int i = 0;
 
         for (Integer integer : integers) {
-            intArray[i++] = integer.intValue();
+            intArray[i++] = integer;
         }
         return intArray;
     }
